requirements.txt
requests
fastapi
langchain
python-dotenv
python-multipart
langsmith